refactor(cli): split the command surface into a package
Summary:
- move the single cli.py into cli/{__init__,source,spec,analyze,schedule,
  inspect}.py, with the parser and dispatch in __init__ and the shared
  source loading in its own module so a command module has no import cycle
- retarget the three schedule-budget patch sites, since the seam followed
  run_schedule into cli/schedule.py
