+++ src/tilefoundry/cli/__main__.py
@@ -1,0 +1,7 @@
+"""``python -m tilefoundry.cli``, the invocation an editable checkout uses."""
+
+from __future__ import annotations
+
+from tilefoundry.cli import main
+
+raise SystemExit(main())
